Enable predicate pushdown for Parquet files read via Hive profile
