+++ server/pxf-hive/src/main/java/org/greenplum/pxf/plugins/hive/HiveAccessor.java
@@ -20,10 +20,15 @@ package org.greenplum.pxf.plugins.hive;
  */
 
 import com.esotericsoftware.kryo.io.Input;
+import com.esotericsoftware.kryo.Kryo;
+import com.esotericsoftware.kryo.io.Output;
+import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang.StringUtils;
 import org.apache.hadoop.hive.common.type.HiveDecimal;
 import org.apache.hadoop.hive.ql.io.IOConstants;
 import org.apache.hadoop.hive.ql.io.orc.Reader;
+import org.apache.hadoop.hive.ql.io.sarg.ConvertAstToSearchArg;
+import org.apache.hadoop.hive.ql.io.sarg.SearchArgument;
 import org.apache.hadoop.hive.serde.serdeConstants;
 import org.apache.hadoop.mapred.InputFormat;
 import org.apache.hadoop.mapred.InputSplit;
@@ -36,8 +41,10 @@ import org.greenplum.pxf.api.filter.Node;
 import org.greenplum.pxf.api.filter.OperandNode;
 import org.greenplum.pxf.api.filter.Operator;
 import org.greenplum.pxf.api.filter.OperatorNode;
+import org.greenplum.pxf.api.filter.SupportedOperatorPruner;
 import org.greenplum.pxf.api.filter.ToStringTreeVisitor;
 import org.greenplum.pxf.api.filter.TreeTraverser;
+import org.greenplum.pxf.api.filter.TreeVisitor;
 import org.greenplum.pxf.api.model.RequestContext;
 import org.greenplum.pxf.api.utilities.ColumnDescriptor;
 import org.greenplum.pxf.plugins.hdfs.HdfsSplittableDataAccessor;
@@ -50,6 +57,7 @@ import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.EnumSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
@@ -91,6 +99,31 @@ public class HiveAccessor extends HdfsSplittableDataAccessor {
     protected List<Integer> hiveIndexes;
     private String hiveColumnsString;
     private String hiveColumnTypesString;
+
+    // ----- members for predicate pushdown handling -----
+    protected Boolean filterInFragmenter;
+
+    private static final int KRYO_BUFFER_SIZE = 4 * 1024;
+    private static final int KRYO_MAX_BUFFER_SIZE = 10 * 1024 * 1024;
+
+    static final EnumSet<Operator> SUPPORTED_OPERATORS =
+            EnumSet.of(
+                    Operator.NOOP,
+                    Operator.LESS_THAN,
+                    Operator.GREATER_THAN,
+                    Operator.LESS_THAN_OR_EQUAL,
+                    Operator.GREATER_THAN_OR_EQUAL,
+                    Operator.EQUALS,
+                    Operator.NOT_EQUALS,
+                    Operator.IS_NULL,
+                    Operator.IS_NOT_NULL,
+                    Operator.IN,
+                    Operator.OR,
+                    Operator.AND,
+                    Operator.NOT
+            );
+    private static final TreeVisitor PRUNER = new SupportedOperatorPruner(SUPPORTED_OPERATORS);
+    private static final TreeTraverser TRAVERSER = new TreeTraverser();
 
     /**
      * Constructs a HiveAccessor
@@ -148,7 +181,8 @@ public class HiveAccessor extends HdfsSplittableDataAccessor {
     }
 
     /**
-     * Opens Hive data split for read. Enables Hive partition filtering. <br>
+     * Opens Hive data split for read. Enables Hive partition filtering.
+     * Adds column projection and predicate pushdown information to JobConf to be used by the reader, if supported.
      *
      * @return true if there are no partitions or there is no partition filter
      * or partition filter is set and the file currently opened by the
@@ -165,8 +199,12 @@ public class HiveAccessor extends HdfsSplittableDataAccessor {
         if (!shouldDataBeReturnedFromFilteredPartition()) {
             return false;
         }
-        // add projected columns to JobConf to make it available to RecordReaders
-        addColumns();
+        if (shouldAddProjectionsAndFilters()) {
+            // add projected columns to JobConf to make them available to RecordReaders
+            addColumns();
+            // add predicate pushdown filters to JobConf to make them available to RecordReaders
+            addFilters();
+        }
         return super.openForRead();
     }
 
@@ -231,6 +269,15 @@ public class HiveAccessor extends HdfsSplittableDataAccessor {
         }
         return inputFormat.getRecordReader(split, jobConf, Reporter.NULL);
     }
+
+    /**
+     * Specifies whether column projection and predicate pushdown information should be added to
+     * the JobConfig so that it is accessible to the RecordReader.
+     * @return true if CP / PPD information should be added, false if not
+     */
+    protected boolean shouldAddProjectionsAndFilters() {
+        return true;
+    };
 
     /*
      * The partition fields are initialized one time base on userData provided
@@ -496,6 +543,49 @@ public class HiveAccessor extends HdfsSplittableDataAccessor {
         jobConf.set(READ_ALL_COLUMNS, "false");
         jobConf.set(READ_COLUMN_IDS_CONF_STR, StringUtils.join(colIds, ","));
         jobConf.set(READ_COLUMN_NAMES_CONF_STR, StringUtils.join(colNames, ","));
+    }
+
+    /**
+     * Uses {@link HiveSearchArgumentBuilder} to translate a filter string into a
+     * Hive {@link SearchArgument} object. The result is added as a filter to
+     * JobConf object
+     */
+    private void addFilters() throws Exception {
+        if (!context.hasFilter()) {
+            return;
+        }
+
+        /* Predicate push-down configuration */
+        String filterStr = context.getFilterString();
+
+        HiveSearchArgumentBuilder searchArgumentBuilder =
+                new HiveSearchArgumentBuilder(context.getTupleDescription(), configuration);
+
+        // Parse the filter string into a expression tree Node
+        Node root = new FilterParser().parse(filterStr);
+        // Prune the parsed tree with valid supported operators and then
+        // traverse the pruned tree with the searchArgumentBuilder to produce a SearchArgument for ORC
+        TRAVERSER.traverse(root, PRUNER, searchArgumentBuilder);
+
+        SearchArgument.Builder filterBuilder = searchArgumentBuilder.getFilterBuilder();
+        SearchArgument searchArgument = filterBuilder.build();
+        jobConf.set(ConvertAstToSearchArg.SARG_PUSHDOWN, toKryo(searchArgument));
+    }
+
+    /**
+     * Package private for unit testing
+     *
+     * @return the jobConf
+     */
+    JobConf getJobConf() {
+        return jobConf;
+    }
+
+    private String toKryo(SearchArgument sarg) {
+        Output out = new Output(KRYO_BUFFER_SIZE, KRYO_MAX_BUFFER_SIZE);
+        new Kryo().writeObject(out, sarg);
+        out.close();
+        return Base64.encodeBase64String(out.toBytes());
     }
 
     protected Properties getSerdeProperties(byte[] userData) {
